spaceDash

+++ components/testimonials.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 interface Props {
-    id: number;
+    id: string;
     spaceId: number;
     name: string;
     email: string;
@@ -10,24 +10,30 @@ interface Props {
     rating: number;
     show: boolean;
 }
-const deleteTestimonial = async (id: number) => {
+const deleteTestimonial = async (id: string) => {
+    const isConfirmed = window.confirm(
+      "Are you sure you want to delete this testimonial?"
+    );
+    if (!isConfirmed) return;
     try {
-        const response = await fetch(`/api/user/deleteTestimonial/${id}`, {
+        console.log(id);
+        await fetch(`/api/user/deleteTestimonial/${id}`, {
             method: "DELETE",
         });
-        console.log(response);
+        window.location.reload();
     } catch (error) {
         console.error("Error deleting testimonial:", error);
     }
 }
 
 
-const updateTestimonial = async (id: number) => {
+const updateTestimonial = async (id: string) => {
     try {
-        const response = await fetch(`/api/user/testimonials/show/${id}`, {
-            method: "POST",
+        await fetch(`/api/user/show/${id}`, {
+          method: "POST",
         });
-        console.log(response);
+        
+        window.location.reload();
     } catch (error) {
         console.error("Error updating testimonial:", error);
     }
@@ -35,27 +41,28 @@ const updateTestimonial = async (id: number) => {
 
 export default function Testimonial({ id, spaceId, name, email, content, videoUrl, rating, show }: Props) {
     return (
-        <div className="p-4 bg-gray-800 rounded-lg shadow-md w-[400px] mt-4">
-            
-                <div>
-                    <h2 className="text-xl font-bold text-white">{name}</h2>
-                    <p className="text-sm text-gray-400">{email}</p>
-                    <p className="mt-2 text-white">{content}</p>
-                    <p className="mt-2 text-yellow-500">Rating: {rating}/5</p>
-                    <button
-                        onClick={() => deleteTestimonial(id)}
-                        className="text-red-500 hover:text-red-700"
-                    >
-                        delete
-                    </button>
-                    <button
-                        onClick={() => updateTestimonial(id)}
-                        className="text-red-500 hover:text-red-700"
-                    >
-                        show
-                    </button>
-                </div>
-            
+      <div className="p-4 bg-gray-800 rounded-lg shadow-md w-full mt-4">
+        <div>
+          <h2 className="text-xl font-bold text-white">{name}</h2>
+          <p className="text-sm text-gray-400">{email}</p>
+          <p className="mt-2 text-white">{content}</p>
+          <p className="mt-2 text-yellow-500">Rating: {rating}/5</p>
+
+          <div className="flex justify-between mt-2">
+            <button
+              onClick={() => deleteTestimonial(id)}
+              className="text-white hover:text-black bg-red-900 items-center px-4 py-1 rounded-md"
+            >
+              delete
+            </button>
+            <button
+              onClick={() => updateTestimonial(id)}
+              className="text-white hover:text-black bg-purple-600 items-center px-4 py-1 rounded-md"
+            >
+              <p>{show ? "Hide" : "Show"}</p>
+            </button>
+          </div>
         </div>
+      </div>
     );
 }
